requirements.txt
# Core RAG
chromadb>=0.4.22
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
tiktoken>=0.5.2

# Document Loading
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
requests>=2.31.0
markdown>=3.5.0

# LLM
groq>=0.4.0

# Observability
langfuse>=2.27.0

# Evaluation
ragas>=0.1.10
datasets>=2.14.0
langchain-groq>=0.1.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiosqlite>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
